tests/wpt/web-platform-tests/css/cssom/getComputedStyle-layout-dependent-replaced-into-ib-split.html

<!doctype html>
<title>getComputedStyle() returns the right style for layout-dependent properties for nodes that have been just inserted into the document, and that have an ancestor whose layout tree was recreated (like an IB-split)</title>
<link rel="help" href="https://drafts.csswg.org/cssom/#dom-window-getcomputedstyle">
<link rel="help" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1585882">
<link rel="author" title="Emilio Cobos Ãlvarez" href="mailto:emilio@crisal.io">
<link rel="author" title="Mozilla" href="https://mozilla.org">
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<style>
  div {
    width: 100%;
  }
</style>
<span>
  <div></div>
</span>
<script>
test(() => {
  let oldDiv = document.querySelector("div");
  window.unused = oldDiv.getBoundingClientRect(); // update layout

  let oldWidth = getComputedStyle(oldDiv).width;
  assert_true(oldWidth.indexOf("px") !== -1, "Should return the used value for width");

  let newDiv = document.createElement("div");
  oldDiv.replaceWith(newDiv);

  assert_equals(getComputedStyle(newDiv).width, oldWidth, "Should return the used value for width (just inserted into the document)");
}, "getComputedStyle() should return used value correctly for nodes just inserted into the document, even if they're in an IB-split");
</script>
